requirements.txt
selene == 2.0.0b17
Appium-Python-Client==2.8.1
pytest==7.2.2
pytest-xdist==3.2.0
allure-pytest==2.13.0
requests==2.28.2
pydantic==1.10.6
python-dotenv==1.0.0